hwmon: (scmi-hwmon) Match scmi device by both name and protocol id
[ Upstream commit b017b2791c9cd6b2e6ec28265e7da3069f25c34f ]

The scmi bus now has support to match the driver with devices not only
based on their protocol id but also based on their device name if one is
available. This was added to cater the need to support multiple devices
and drivers for the same protocol.

Let us add the name "hwmon" to scmi_device_id table in the driver so
that in matches only with device with the same name and protocol id
SCMI_PROTOCOL_SENSOR. This is just for sake of completion and must
not be used to add IIO support in parallel. Instead, if IIO support is
added ever in future, we need to drop this hwmon driver entirely and
use the iio->hwmon bridge to access the sensors as hwmon devices if
needed.

Cc: linux-hwmon@vger.kernel.org
Acked-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>




[mgo@opensynergy.com: cherry-picked from: 
repo: https://github.com/OpenSynergy/android-kernel-common.git
branch: opsy/android11-5.4-trout
commit: 235b2b73e686b407df1d5a40cce2ab40a0103be2]

Issue: COQOS-14841
Change-Id: I49f4f6213e0e01cb4b491f2bd2a7ad401808ee47