drivers: media: Apply missing changes from OpSy's driver published on Github.
The QTI driver contains extra/missing changes which were not accounted
for and unfortunately do not work with coqos-virtio-video - every
decoded frame is black. This patch is based on a diff between the
version OpSy published[1] and the version QTI published[2].

The main offending issue is probably the order in which commits were
imported. In the original publication the order of patches is:
  $git log --oneline ./drivers/media/virtio
  1f067e9fb media: virtio: Update driver version string to: v2.3.0.
  d46bfe1b4 media: virtio: Move VIRTIO_IDs from driver's header to uapi header.
  483ae2f57 media: virtio: Import source code of coqos virtio video driver v2.3.0.
  b6d9bad39 media: Add Kconfig and Makefile for virtio-video module.
  81afdb26e media: Import virtio-video device driver source code.

Whereas in the target version the order is:
  855e711bfece media: Import virtio-video device driver source code.
  13fc89c75ee1 media: virtio: Update driver version string to: v2.3.0.
  afb5a9ae6572 media: virtio: Move VIRTIO_IDs from driver's header to uapi header.
  d7af16f5b389 media: virtio: Import source code of coqos virtio video driver v2.3.0.
  c1f28295a857 media: Add Kconfig and Makefile for virtio-video module.

As one might suspect the commits "Add Kconfig and Makefile for
virtio-video module" and "media: virtio: Import source code of coqos
virtio video driver v2.3.0" in the QTI driver do not match the
originally published changes.

Unfortunately due to the change in order in which patches were applied
the diff between the versions is not 0 lines. This patch contains
exactly the missing diff.

[1] https://github.com/OpenSynergy/android-kernel-common/tree/1f067e9fb8cb3c156fe4102ccfa9174eddeb5aa8/drivers/media/virtio
[2] https://git.codelinaro.org/clo/la/kernel/msm-5.4/-/tree/855e711bfecea94e1a7ba9aa7d1798e69c561d03/drivers/media/virtio

Issue: COQOS-14749, E0927-1512, QCOM-250
Change-Id: Id2825c4d5bcda7f391915418b972f64a47adcca5