media: virtio: Remove duplicated definitions from virtio_video.h.
These definitions exists in driver's origin repository but are
also defined in uapi header after integration the driver v2.3.0:
* include/uapi/linux/virtio_ids.h

Remove it from imported driver header to avoid the build break.

Note: This is a workaround. Until a proper fix is available,
the internal and the public version of 'virtio_video.h' are
out-of-sync w.r.t. this change.

Issue: COQOS-17648
Change-Id: I18759c43ef244287e360adc41f2a43cd3358b742